Divicion de banner
algunos alementos fueron enviados a JIN26/Banner

+++ index.sh
@@ -1,43 +1,34 @@
 #!/bin/bash
 
-function mostrar {
-    echo "Code: $code"
-    curl ifconfig.me||cowsay sorry
-    echo
-    cd ini
-    chmod 777 vnc.sh
-    ./vnc.sh &
-    cd ..
-}
+PWDJIN=$PWD;#Direccion del proyecto
 
 #Actualizacion
-sudo apt update && apt upgrade;
+#sudo apt update && sudo apt upgrade;
 
 #Jin
-PWDJIN=$PWD;
+if [ -e ~/.Jin ]; then
 mkdir ~/.Jin
 cd ~/.Jin
+fi
+#logo
+if [ -e ~/.Jin/logo ]; then
 sudo ln -s $PWDJIN/img/logo.sh logo
 chmod 777 logo
 sudo cp logo $PREFIX/bin
 sudo cp logo $PREFIX/bin/applets
-
+fi
 #Servidor localhost
+if [ -e ~/.Jin/index ]; then
 sudo ln -s $PWDJIN/index.php index
 chmod 777 index
 sudo cp index ~/Desktop
-
-#ip php
-sudo ln -s $PWDJIN/php/ip.php ip
-chmod 777 ip
+fi
 
 logo
 cd $PWDJIN
-
-read -p "Que tipo de lenguaje quieres usar $ " code
-#if[$code="123"];then
-    mostrar
-#fi
+cd ini
+chmod 777 vnc.sh
+./vnc.sh &
 
 # ssh llave install.
 #ssh-keygen -t rsa -b 4096 -C "Angel_mantilla_26@hotmail.com"
@@ -45,25 +36,23 @@ eval $(ssh-agent -s)
 #ssh-add ~/.ssh/id_rsa
 
 #git
-sh ini/git.sh
+sh git.sh
 
 #editor
-sh ini/nvim.sh
+sh nvim.sh
 
 case "$OSTYPE" in
     cygwin*)
         ;;
     linux-androideabi)
         sleep 2;
-        cd ini;
-	chmod 777 termux.sh
+	    chmod 777 termux.sh
         sh termux.sh
         ;;
     linux-gnu)
         sleep 2;
-        cd ini;
         chmod 777 linux.sh
-        sh linux.sh
+        #sh linux.sh
         ;;
     darwin*)
         ;;
